+++ firmware/config/boards/hellen/hellen128/board_configuration.cpp
@@ -117,7 +117,6 @@ void setBoardConfigOverrides(void) {
 
 	engineConfiguration->canTxPin = GPIOD_1;
 	engineConfiguration->canRxPin = GPIOD_0;
-	hellenWbo();
 }
 
 void setPinConfigurationOverrides(void) {
@@ -171,6 +170,8 @@ void setBoardDefaultConfiguration(void) {
 
 	strcpy(CONFIG(engineMake), ENGINE_MAKE_MERCEDES);
 	strcpy(CONFIG(engineCode), "");
+
+	hellenWbo();
 }
 
 /**