+++ firmware/config/boards/subaru_eg33/config/controllers/algo/auto_generated_enums.cpp
@@ -1,7 +1,7 @@
 #include "global.h"
 #include "rusefi_enums.h"
 #include "rusefi_hw_enums.h"
-// was generated automatically by rusEFI tool  from rusefi_hw_enums.h // was generated automatically by rusEFI tool  from rusefi_enums.h // by enum2string.jar tool on Thu Jul 08 03:47:45 UTC 2021
+// was generated automatically by rusEFI tool  from rusefi_hw_enums.h // was generated automatically by rusEFI tool  from rusefi_enums.h // by enum2string.jar tool on Thu Jul 08 20:15:18 EDT 2021
 // see also gen_config_and_enums.bat
 
 
@@ -898,10 +898,12 @@ case HELLEN_121_VAG:
   return "HELLEN_121_VAG";
 case HELLEN_121_VAG_5_CYL:
   return "HELLEN_121_VAG_5_CYL";
-case HELLEN_121_VAG_6_CYL:
-  return "HELLEN_121_VAG_6_CYL";
 case HELLEN_121_VAG_8_CYL:
   return "HELLEN_121_VAG_8_CYL";
+case HELLEN_121_VAG_V6_CYL:
+  return "HELLEN_121_VAG_V6_CYL";
+case HELLEN_121_VAG_VR6_CYL:
+  return "HELLEN_121_VAG_VR6_CYL";
 case HELLEN_128_MERCEDES:
   return "HELLEN_128_MERCEDES";
 case HELLEN_134_BMW: