+++ firmware/config/engines/vw.cpp
@@ -73,10 +73,18 @@ void setHellen121Vag_5_cyl(DECLARE_CONFIG_PARAMETER_SIGNATURE) {
 	engineConfiguration->specs.firingOrder = FO_1_2_4_5_3;
 }
 
-void setHellen121Vag_6_cyl(DECLARE_CONFIG_PARAMETER_SIGNATURE) {
+void setHellen121Vag_vr6_cyl(DECLARE_CONFIG_PARAMETER_SIGNATURE) {
+	engineConfiguration->specs.cylindersCount = 6;
+	engineConfiguration->specs.firingOrder = FO_1_5_3_6_2_4;
+}
+
+void setHellen121Vag_v6_cyl(DECLARE_CONFIG_PARAMETER_SIGNATURE) {
 	engineConfiguration->specs.cylindersCount = 6;
 	engineConfiguration->specs.displacement = 2.7;
-// todo? typo? missing?	engineConfiguration->specs.firingOrder = FO_1_4_3_6_2_5;
+
+	engineConfiguration->camInputs[1 * CAMS_PER_BANK] = GPIOA_7; // 87a
+
+	engineConfiguration->specs.firingOrder = FO_1_4_3_6_2_5;
 }
 
 void setHellen121Vag_8_cyl(DECLARE_CONFIG_PARAMETER_SIGNATURE) {