+++ java_tools/configuration_definition/src/com/rusefi/test/ConfigFieldParserTest.java
@@ -81,12 +81,23 @@ public class ConfigFieldParserTest {
         new ReaderState().readBufferedReader(reader, Collections.emptyList());
     }
 
+    @Test(expected = IllegalStateException.class)
+    public void invalidDefine() throws IOException {
+        String test = "struct pid_s\n" +
+                ReaderState.DEFINE + " show show_Hellen121vag_presets true\n" +
+                "end_struct\n" +
+                "";
+        VariableRegistry.INSTANCE.clear();
+        BufferedReader reader = new BufferedReader(new StringReader(test));
+        new ReaderState().readBufferedReader(reader, Collections.emptyList());
+    }
+
     @Test
     public void multiplicationInDefine() throws IOException {
         String test = "struct pid_s\n" +
                 "#define ERROR_BUFFER_SIZE 120\n" +
                 "#define ERROR_BUFFER_COUNT 120\n" +
-                "#define RESULT @@ERROR_BUFFER_SIZE@@ * @@ERROR_BUFFER_COUNT@@\n" +
+                "#define RESULT @@ERROR_BUFFER_SIZE@@*@@ERROR_BUFFER_COUNT@@\n" +
                 "\tint16_t periodMs;PID dTime;\"ms\",      1,      0,       0, 3000,      0\n" +
                 "end_struct\n" +
                 "";